Changed to uint64 for hash coding in sparese matrix pattern.

+++ externals/IBKMK/src/IBKMK_SparseMatrixPattern.h
@@ -109,10 +109,10 @@ public:
 	\param endIndex end index of matrix part whose transitive closure should be calculated.
 	*/
 	static void calculateTransitiveClosure(
-		SparseMatrixPattern &pattern, 
-		SparseMatrixPattern &transposePattern, 
+		SparseMatrixPattern &pattern,
+		SparseMatrixPattern &transposePattern,
 		unsigned int n,
-		unsigned int startIndex, 
+		unsigned int startIndex,
 		unsigned int endIndex);
 
 	/*! Default constructor, registers size of index space.
@@ -147,7 +147,7 @@ private:
 #ifdef STANDARD_IMPLEMENTATION
 	std::set<unsigned int>	m_data;
 #else // STANDARD_IMPLEMENTATION
-	std::map<unsigned int, unsigned int>	m_data;
+	std::map<uint64_t, unsigned int>	m_data;
 #endif // STANDARD_IMPLEMENTATION
 };
